# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns access token, or a temporary token when 2FA is required",
            "content": {
              "application/json": {
                "examples": {
                  "session": {
                    "summary": "JWT issued",
                    "value": {
                      "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example",
                      "user": {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "email": "trader@example.com",
                        "displayName": "Demo Trader",
                        "role": "user",
                        "emailVerifiedAt": "2026-04-09T12:00:00.000Z",
                        "createdAt": "2026-04-09T12:00:00.000Z",
                        "updatedAt": "2026-04-09T12:00:00.000Z",
                        "profile": {
                          "username": "demotrader",
                          "fullName": "Demo Trader"
                        }
                      }
                    }
                  },
                  "twoFactor": {
                    "summary": "2FA verification required",
                    "value": {
                      "requires2FA": true,
                      "tempToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.temp2fa"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "summary": "Request password reset code",
        "description": "Sends an 8-digit code to the email when the account exists. Response is always the same to avoid email enumeration.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generic confirmation message (same whether or not the email exists)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "If an account exists for this email, a reset code has been sent. Check your inbox."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "summary": "Reset password with email and code",
        "description": "Consumes the code from the reset email and sets a new password. Invalidates all sessions for the user.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordWithCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Invalid or expired reset code",
                    "error": "Bad Request"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns user and access token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example",
                    "user": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "trader@example.com",
                      "displayName": "Demo Trader",
                      "role": "user",
                      "emailVerifiedAt": "2026-04-09T12:00:00.000Z",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "updatedAt": "2026-04-09T12:00:00.000Z",
                      "profile": {
                        "username": "demotrader",
                        "fullName": "Demo Trader"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/admin/register": {
      "post": {
        "operationId": "AuthController_createAdmin",
        "summary": "Create admin user (protected by ADMIN_API_KEY)",
        "description": "Creates an admin user. Requires X-Admin-API-Key header (or Authorization: Bearer <key>) matching ADMIN_API_KEY in .env.",
        "parameters": [
          {
            "name": "X-Admin-API-Key",
            "in": "header",
            "description": "Admin API key from ADMIN_API_KEY env",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns admin user and access token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.admin",
                    "user": {
                      "id": "660e8400-e29b-41d4-a716-446655440001",
                      "email": "admin@example.com",
                      "displayName": "Admin",
                      "role": "admin",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "updatedAt": "2026-04-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid or missing admin API key",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 401,
                    "message": "Invalid or missing admin API key",
                    "error": "Unauthorized"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/admin/create-user": {
      "post": {
        "operationId": "AuthController_createUserAsAdmin",
        "summary": "Create user with profile (admin only)",
        "description": "Persists user and optional profile to the database (same rules as register-with-profile). Does not create a session for the new user.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created (no session for the new user)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "550e8400-e29b-41d4-a716-446655440000",
                    "email": "trader@example.com",
                    "displayName": "Demo Trader",
                    "role": "user",
                    "profile": {
                      "username": "demotrader",
                      "fullName": "Demo Trader",
                      "photoUrl": null,
                      "bio": "QA practice account",
                      "websiteUrl": null,
                      "location": "EU",
                      "birthday": null,
                      "languageCode": "en",
                      "timezone": "UTC",
                      "preferences": {}
                    },
                    "createdAt": "2026-04-09T12:00:00.000Z",
                    "updatedAt": "2026-04-09T12:00:00.000Z"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email or username conflict",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 409,
                    "message": "Email already registered",
                    "error": "Conflict"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/admin/bulk-import-users": {
      "post": {
        "operationId": "AuthController_bulkImportUsers",
        "summary": "Bulk import users from CSV or JSON (admin only)",
        "description": "Single multipart upload. Rows are validated like the import template; duplicates in-DB yield skipped; invalid rows yield error entries. Response lists per-row outcomes in file order.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV or JSON (max 500 rows, 5 MB). Same columns as single create-user."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created / failed / skipped counts and per-row outcomes",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "created": 2,
                    "failed": 0,
                    "skipped": 1,
                    "rows": [
                      {
                        "email": "new@example.com",
                        "status": "created",
                        "userId": "880e8400-e29b-41d4-a716-446655440003"
                      },
                      {
                        "email": "dup@example.com",
                        "status": "skipped",
                        "message": "Email already registered"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing file or parse error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 400,
                    "message": "Invalid or expired reset code",
                    "error": "Bad Request"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/register-with-profile": {
      "post": {
        "operationId": "AuthController_registerWithProfile",
        "summary": "Register new user with optional profile fields",
        "description": "Creates a user (same as register) plus optional UserProfile fields in one request.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWithProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns user and access token",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example",
                    "user": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "trader@example.com",
                      "displayName": "Demo Trader",
                      "role": "user",
                      "emailVerifiedAt": "2026-04-09T12:00:00.000Z",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "updatedAt": "2026-04-09T12:00:00.000Z",
                      "profile": {
                        "username": "demotrader",
                        "fullName": "Demo Trader"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/impersonate": {
      "post": {
        "operationId": "AuthController_impersonate",
        "summary": "Impersonate a user (admin only)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns target user token and backToAdminToken",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.imp",
                    "backToAdminToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.back",
                    "user": {
                      "id": "770e8400-e29b-41d4-a716-446655440002",
                      "email": "target@example.com",
                      "displayName": "Target User",
                      "role": "user",
                      "impersonatedBy": "660e8400-e29b-41d4-a716-446655440001"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 403,
                    "message": "Admin access required",
                    "error": "Forbidden"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/end-impersonation": {
      "post": {
        "operationId": "AuthController_endImpersonation",
        "summary": "End impersonation and return to admin account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EndImpersonationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns admin session (access token + user)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.admin",
                    "user": {
                      "id": "660e8400-e29b-41d4-a716-446655440001",
                      "email": "admin@example.com",
                      "displayName": "Admin",
                      "role": "admin",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "updatedAt": "2026-04-09T12:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Logout (invalidate session)",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session invalidated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/2fa/verify": {
      "post": {
        "operationId": "AuthController_verify2Fa",
        "summary": "Verify 2FA during login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify2FaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns access token after 2FA verification",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example",
                    "user": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "email": "trader@example.com",
                      "displayName": "Demo Trader",
                      "role": "user",
                      "emailVerifiedAt": "2026-04-09T12:00:00.000Z",
                      "createdAt": "2026-04-09T12:00:00.000Z",
                      "updatedAt": "2026-04-09T12:00:00.000Z",
                      "profile": {
                        "username": "demotrader",
                        "fullName": "Demo Trader"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/status": {
      "get": {
        "operationId": "AuthController_get2FaStatus",
        "summary": "Get 2FA status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns { enabled: boolean }",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "enabled": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/auth/2fa/backup-codes": {
      "get": {
        "operationId": "AuthController_getBackupCodes",
        "summary": "Get backup codes (throws - use regenerate)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Original codes are not retrievable after setup",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "codes": [],
                    "message": "Backup codes are shown only once during 2FA setup. Use regenerate to get new codes."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Test Exchange API",
    "description": "Crypto exchange training platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "you@example.com",
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordWithCodeDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "you@example.com",
            "maxLength": 254
          },
          "code": {
            "type": "string",
            "description": "8-digit code from the reset email (spaces allowed)"
          },
          "newPassword": {
            "type": "string",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "code",
          "newPassword"
        ]
      },
...
    }
  }
}